Move color conversion functions to utilities
